[CPCHawk] Some cleanup
