+++ src/BizHawk.Emulation.Cores/Computers/AmstradCPC/Machine/CPC6128/CPC6128.cs
@@ -16,12 +16,11 @@ namespace BizHawk.Emulation.Cores.Computers.AmstradCPC
 			CPC = cpc;
 			CPU = cpu;
 
-			FrameLength = 79872;
-
 			CRCT = new CRTC(0);
-			//CRCT = new CRCT_6845(CRCT_6845.CRCTType.MC6845, this);
-			//CRT = new CRTDevice(this);
-			GateArray = new GateArray(this, AmstradGateArrayType.Amstrad40010);
+			GateArray = new GateArray(this, GateArrayType.Amstrad40010);
+
+			FrameLength = GateArray.FrameLength / 4;
+
 			PPI = new PPI_8255(this);
 
 			TapeBuzzer = new Beeper(this);
